pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "whiterabbit"
version = "0.1.0"
description = "Secure file laundering and obfuscation library"
authors = [
  { name="Miscreant", email="aMiscreant@protonmail.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "Pillow", "piexif", "numpy", "cryptography", "requests", "flask"
]

[project.scripts]
whiterabbit = "whiterabbit.main:main"

[tool.setuptools.packages.find]
where = ["whiterabbit"]
